Add 404 error pages and improve user management scripts; implement error handling and session checks

+++ admin/script/delete-user.php
@@ -1,13 +1,19 @@
 <?php 
 
     include "config.php";
-    $DELETE_USER_ID = $_POST['UserId'];
-    $query = "DELETE FROM users WHERE id = {$DELETE_USER_ID}";
-    $result = mysqli_query($conn, $query);
-    if($result){
-        echo "User deleted";
+    if($_SERVER['REQUEST_METHOD'] === 'POST'){
+
+        $DELETE_USER_ID = $_POST['UserId'];
+        $query = "DELETE FROM users WHERE id = {$DELETE_USER_ID}";
+        $result = mysqli_query($conn, $query);
+        if($result){
+            echo "User deleted";
+        }else{
+            echo mysqli_error($conn);
+        }
     }else{
-        echo mysqli_error($conn);
+        header("Location: {$host_name}/admin/not-found.php");
+        exit();
     }
 
 ?>
